core: inspect: move kernel specifics into a KernelInspector
